Create main.py
The purpose of main.py is to read stock data from a JSON file, store it in a SQLite database, retrieve the data from the database, and plot the stock data using the plot_stock_data function. It organizes the flow of the program and utilizes the functionalities provided by other modules to accomplish this task.